resumeParsing/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
chromadb>=0.5.3
sentence-transformers>=2.5.1
python-multipart>=0.0.9
pypdf>=4.2.0
python-docx>=1.1.2
pydantic>=2.7.1


